refactor: Move the sentence graph class to new file, and fix multi_config active list
