+++ multilingual_ime/key_event_handler.py
@@ -1,5 +1,5 @@
 import logging
-import heapq
+
 from pathlib import Path
 
 from .candidate import Candidate
@@ -15,7 +15,7 @@ from .ime import (
 )
 from .phrase_db import PhraseDataBase
 from .muti_config import MultiConfig
-
+from .sentence_graph import SentenceGraph
 
 from .ime import (
     BOPOMOFO_VALID_KEYSTROKE_SET,
@@ -53,7 +53,6 @@ class KeyEventHandler:
         self._user_frequency_db = KeystrokeMappingDB(USER_FREQUENCY_DB_PATH)
 
         # Setup IMEs
-        self.activated_imes: list[str] = self._config.ACTIVE_IME
         self.ime_handlers = {
             ime: IMEFactory.create_ime(ime) for ime in self.activated_imes
         }
@@ -131,8 +130,11 @@ class KeyEventHandler:
 
     def set_activation_status(self, ime_type: str, status: bool) -> None:
         self._config.setIMEActivationStatus(ime_name=ime_type, status=status)
-        self.activated_imes = self._config.ACTIVE_IME
 
+    @property
+    def activated_imes(self) -> list[str]:
+        return self._config.ACTIVE_IME
+        
     @property
     def token_pool(self) -> list[str]:
         return list(self._token_pool_set)
@@ -612,80 +614,6 @@ class KeyEventHandler:
         raise NotImplementedError("update_user_phrase_db is not implemented yet")
 
     def new_reconstruct(self, keystroke: str, top_n: int = 10) -> list[list[str]]:
-        class SentenceGraph:
-            def __init__(self) -> None:
-                self._graph = {}
-
-            @property
-            def num_of_node(self) -> int:
-                return len(self._graph)
-
-            def add_edge(
-                self, u_id: str, v_id: str, distance: int, direct: bool = True
-            ) -> None:
-                if u_id not in self._graph:
-                    self._graph[u_id] = [(v_id, distance)]
-                else:
-                    if (v_id, distance) not in self._graph[u_id]:
-                        self._graph[u_id].append((v_id, distance))
-
-                if v_id not in self._graph:
-                    if direct:
-                        self._graph[v_id] = []
-                    else:
-                        self._graph[v_id] = [(u_id, distance)]
-
-            def find_shortest_paths(self, start_id: str, end_id: str) -> list[str]:
-                # By Dijkstra
-                predecessor = {id: None for id in self._graph}
-                distance = {id: None for id in self._graph}
-                distance[start_id] = 0
-
-                priority_queue = [(0, start_id)]
-                while priority_queue:
-                    current_distance, current_id = heapq.heappop(priority_queue)
-
-                    for neighbor_id, neighbor_weight in self._graph[current_id]:
-                        neg_new_distance = current_distance + neighbor_weight
-
-                        if distance[neighbor_id] is None:
-                            distance[neighbor_id] = neg_new_distance
-                            heapq.heappush(
-                                priority_queue, (neg_new_distance, neighbor_id)
-                            )
-                            predecessor[neighbor_id] = set([current_id])
-                        else:
-
-                            if neg_new_distance < distance[neighbor_id]:
-                                distance[neighbor_id] = neg_new_distance
-                                heapq.heappush(
-                                    priority_queue, (neg_new_distance, neighbor_id)
-                                )
-                                predecessor[neighbor_id] = set([current_id])
-                            elif neg_new_distance == distance[neighbor_id]:
-                                predecessor[neighbor_id].add(current_id)
-
-                # Get the path
-                def get_path(
-                    predecessor: dict[str, set], end_id: str
-                ) -> list[list[str]]:
-
-                    def dfs(current_id: str) -> list[list[str]]:
-                        if current_id == start_id:
-                            return [[start_id]]
-
-                        if predecessor[current_id] is None:
-                            return []
-
-                        paths = []
-                        for pred in predecessor[current_id]:
-                            paths.extend([path + [current_id] for path in dfs(pred)])
-                        return paths
-
-                    return dfs(end_id)
-
-                return get_path(predecessor, end_id)
-
         if not keystroke:
             return []
 
@@ -720,6 +648,8 @@ class KeyEventHandler:
         self.logger.info("Creating Graph with %d possible seps", len(new_possible_seps))
         id_maps = {}
         graph = SentenceGraph()
+
+        # TODO: Hide the creation of the graph in the SentenceGraph class
         for sep in new_possible_seps:
             prev_str = ""
             prev_token_id = "<start>"
